requirements.txt
requests
dnspython
python-whois
colorama
rich
